feat: add the published date to Work
Dates in the Crossref API can be just a month and a year, or even just a year. Returning the
integers themselves isn't particularly useful, nor is trying to create a Date object.

Instead, this change introduces the Temporal proposal and uses a polyfill. While this isn't ideal,
the proposal seems stable for this usage and is better than the alternatives.

Refs https://tc39.es/proposal-temporal/